add
meta 업그레이드 ui 및 관련 json들 불러오기 완료
meta 업그레이드 상호작용 등 완료
stage에서 배속 추가
씬 이동간 로딩 추가 필요
option제작중

+++ Assets/02.Scripts/UI/View/Stage/TowerGradeUpgradeView.cs
@@ -98,7 +98,7 @@ public class TowerGradeUpgradeView : MonoBehaviour, IPopUpPublicUI
     public void SetTowerName(string name) => towerNameText.text = name;
     public void SetSkillName(string name) => towerSkillText.text = name;
     public void SetSkillDes(string des) => towerSkillDesText.text = des;
-    public void SetDamageCurrentValue(int value) => damageCurrentValueText.text = value.ToString();
+    public void SetDamageCurrentValue(float value) => damageCurrentValueText.text = value.ToString();
     public void SetAttackSpeedCurrentValue(float value) => attackSpeedCurrentValueText.text = value.ToString();
     public void SetRangeCurrentValue(float value) => rangeCurretnValueText.text = value.ToString();
     public void PremiumUpgradePirce(int value) => premiunUpgradePriceText.text = value.ToString();
